Add OmimDisease annotations
